refactor(sqlcomment): move package from pkg/db to pkg/mysql
Frontline and other services migrating off pkg/db import pkg/mysql for
database access. Colocating sqlcomment with pkg/mysql keeps annotations
working as pkg/db is phased out. pkg/db continues to depend on the
moved package for legacy callers.
