+++ cmd/dev/seed/checkpoints/BUILD.bazel
@@ -15,8 +15,8 @@ go_library(
         "//pkg/clickhouse",
         "//pkg/clickhouse/schema",
         "//pkg/db",
-        "//pkg/db/sqlcomment",
         "//pkg/logger",
+        "//pkg/mysql/sqlcomment",
         "//pkg/prompt",
         "//pkg/uid",
         "@org_golang_x_term//:term",
