+++ docs/engineering/infra/planetscale/query-insights-tags.mdx
@@ -66,7 +66,7 @@ rows, err := database.RO().QueryContext(ctx, query, args...)
 
 `sqlcomment.Static{}` disables annotation (tests and dev seeds).
 
-Package reference: [`pkg/db/sqlcomment`](../../../../pkg/db/sqlcomment/doc.go).
+Package reference: [`pkg/mysql/sqlcomment`](../../../../pkg/mysql/sqlcomment/doc.go).
 
 ## TypeScript apps
 
@@ -95,7 +95,7 @@ After deploy to staging:
 Local unit tests:
 
 ```bash
-mise exec -- bazel test //pkg/db/sqlcomment:sqlcomment_test
+mise exec -- bazel test //pkg/mysql/sqlcomment:sqlcomment_test
 mise exec -- pnpm --dir=web exec vitest run internal/db/src/sqlcomment.test.ts internal/db/src/commented-pool.test.ts
 ```
