+++ pkg/db/BUILD.bazel
@@ -385,12 +385,12 @@ go_library(
     deps = [
         "//pkg/assert",
         "//pkg/codes",
-        "//pkg/db/sqlcomment",
         "//pkg/db/types",
         "//pkg/fault",
         "//pkg/logger",
         "//pkg/mysql",
         "//pkg/mysql/metrics",
+        "//pkg/mysql/sqlcomment",
         "//pkg/otel/tracing",
         "//pkg/retry",
         "@com_github_go_sql_driver_mysql//:mysql",
