+++ pkg/mysql/sqlcomment/static.go
