+++ pkg/zen/BUILD.bazel
@@ -31,9 +31,9 @@ go_library(
         "//pkg/auth/principal",
         "//pkg/clickhouse/schema",
         "//pkg/codes",
-        "//pkg/db/sqlcomment",
         "//pkg/fault",
         "//pkg/logger",
+        "//pkg/mysql/sqlcomment",
         "//pkg/otel/tracing",
         "//pkg/prometheus/metrics",
         "//pkg/tls",
@@ -71,8 +71,8 @@ go_test(
         "//pkg/auth/principal",
         "//pkg/clickhouse/schema",
         "//pkg/codes",
-        "//pkg/db/sqlcomment",
         "//pkg/fault",
+        "//pkg/mysql/sqlcomment",
         "//pkg/tls",
         "//svc/api/openapi",
         "@com_github_stretchr_testify//assert",
