+++ svc/api/BUILD.bazel
@@ -34,8 +34,8 @@ go_library(
         "//pkg/config",
         "//pkg/counter",
         "//pkg/db",
-        "//pkg/db/sqlcomment",
         "//pkg/logger",
+        "//pkg/mysql/sqlcomment",
         "//pkg/otel",
         "//pkg/prometheus",
         "//pkg/prometheus/lazy",
