+++ svc/ctrl/api/BUILD.bazel
@@ -22,9 +22,9 @@ go_library(
         "//pkg/clickhouse/schema",
         "//pkg/clock",
         "//pkg/config",
-        "//pkg/db/sqlcomment",
         "//pkg/dns/domainconnect",
         "//pkg/logger",
+        "//pkg/mysql/sqlcomment",
         "//pkg/otel",
         "//pkg/prometheus",
         "//pkg/prometheus/lazy",
@@ -66,8 +66,8 @@ go_test(
         "//gen/proto/ctrl/v1/ctrlv1connect",
         "//gen/proto/hydra/v1:hydra",
         "//pkg/config",
-        "//pkg/db/sqlcomment",
         "//pkg/logger",
+        "//pkg/mysql/sqlcomment",
         "//pkg/rpc/interceptor",
         "//pkg/testutil/containers",
         "//pkg/uid",
