+++ svc/ctrl/integration/harness/BUILD.bazel
@@ -13,8 +13,8 @@ go_library(
         "//pkg/clickhouse",
         "//pkg/clickhouse/schema",
         "//pkg/clock",
-        "//pkg/db/sqlcomment",
         "//pkg/healthcheck",
+        "//pkg/mysql/sqlcomment",
         "//pkg/restate/admin",
         "//pkg/testutil/containers",
         "//svc/ctrl/integration/seed",
