+++ svc/ctrl/internal/db/BUILD.bazel
@@ -11,7 +11,7 @@ go_library(
     importpath = "github.com/unkeyed/unkey/svc/ctrl/internal/db",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/db/sqlcomment",
+        "//pkg/mysql/sqlcomment",
         "//pkg/fault",
         "//pkg/mysql",
         "//pkg/mysql/types",
