+++ svc/ctrl/internal/db/database.go
@@ -3,7 +3,7 @@ package db
 import (
 	"context"
 
-	"github.com/unkeyed/unkey/pkg/db/sqlcomment"
+	"github.com/unkeyed/unkey/pkg/mysql/sqlcomment"
 	"github.com/unkeyed/unkey/pkg/fault"
 	"github.com/unkeyed/unkey/pkg/mysql"
 )
