+++ svc/ctrl/worker/BUILD.bazel
@@ -22,9 +22,9 @@ go_library(
         "//pkg/clickhouse/schema",
         "//pkg/clock",
         "//pkg/config",
-        "//pkg/db/sqlcomment",
         "//pkg/healthcheck",
         "//pkg/logger",
+        "//pkg/mysql/sqlcomment",
         "//pkg/otel",
         "//pkg/prometheus",
         "//pkg/prometheus/lazy",
