+++ svc/frontline/BUILD.bazel
@@ -27,8 +27,8 @@ go_library(
         "//pkg/config",
         "//pkg/counter",
         "//pkg/db",
-        "//pkg/db/sqlcomment",
         "//pkg/logger",
+        "//pkg/mysql/sqlcomment",
         "//pkg/otel",
         "//pkg/pprof",
         "//pkg/prometheus",
