+++ svc/frontline/internal/policies/BUILD.bazel
@@ -54,9 +54,9 @@ go_test(
         "//pkg/codes",
         "//pkg/counter",
         "//pkg/db",
-        "//pkg/db/sqlcomment",
         "//pkg/fault",
         "//pkg/hash",
+        "//pkg/mysql/sqlcomment",
         "//pkg/ptr",
         "//pkg/rbac",
         "//pkg/testutil/containers",
